requirements.txt
annotated-types==0.5.0
anyio==3.6.2
bson==0.5.10
bump-pydantic==0.6.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
config==0.5.1
cryptography==41.0.3
DateTime==4.7
dnspython==2.2.1
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.101.0
fastapi-jwt==0.1.12
fastapi-jwt-auth==0.5.0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
libcst==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
motor==3.2.0
mypy-extensions==1.0.0
orjson==3.8.3
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
Pygments==2.15.1
PyJWT==1.7.1
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rich==13.5.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typer==0.9.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.7.1
ujson==5.6.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zope.interface==5.5.2
